Add ability to export a submodel as a custom model ... this allows you to export a submodel. Rework the models node numbers then re-import the custom model. You do lose the buffer layout but the submodelling comes in.

+++ xLights/models/Model.h
@@ -326,7 +326,7 @@ public:
     virtual uint32_t GetLastChannel() const;
     uint32_t GetNumChannels();
     uint32_t GetNodeNumber(size_t nodenum) const;
-    uint32_t GetNodeNumber(int bufY, int bufX) const;
+    long GetNodeNumber(int bufY, int bufX) const;
     bool UpdateStartChannelFromChannelString(std::map<std::string, Model*>& models, std::list<std::string>& used);
     int GetNumberFromChannelString(const std::string& sc) const;
     int GetNumberFromChannelString(const std::string& sc, bool& valid, std::string& dependsonmodel) const;
